+++ CHANGES.rst
@@ -1,3 +1,11 @@
+0.0.2 (2025-06-03)
+----------------
+
+- Added support for optically thick dust.
+- Fixed bug in the plotting limits.
+- Improved readthedocs documentation.
+- Added support for NIRCam.
+
 0.0.1 (2025-05-08)
 ----------------